pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "black>=24.4.2",
    "python-dotenv>=1.0.1",
    "isort>=5.13.2",
    "langchain>=0.2.16",
    "langchain-ollama>=0.1.1",
    "langchain-openai>=0.1.8",
    "google-adk>=1.21.0",
    "openai>=2.11.0",
    "ollama>=0.6.1",
    "plantuml>=0.3.0",
]

